Health Bars
Got a really simple GUI based health bar done, However! this way seems
to be unefficient and needs to be fixed. part of the reason as to why
it's bad is because I'm having diffuculty placing P2 health bar in the
right spot. I will try to find a better way to do this tomorrow